requirements.txt
python-osc>=1.8.0
sounddevice>=0.4.0
numpy>=1.21.0
soundfile>=0.12.0
openai-whisper>=20231117